Update make, use eslint and other QOL changes
Switches linting to eslint, preserving most of the rules set-up for tslint before, but allowing for 2 different configurations:

- common holds student configuration. This is what they pull down, and if their IDE supports it, or if they run yarn lint, this is what is used.
- maintaining holds track configuration/configuration for maintainers. This is stricter and used by the ci and make file.

The configurations have been loosened quite a bit for students, in order to only catch bugs, but not dictate any style. Before, there were quite a few stylistic choices in tslint set-up.

The identation has been normalised. A big chunk of the files was indented 2 spaces and another chunk 4 spaces. Sometimes it was only the tests and sometimes only the example. I chose the rule that matches javascript, which uses 2 spaces. In the next commit you'll see the update.

Updated the make file to run the lintreport from eslint, using the track package, as well as updating them for cross-env:

- Use quotes around file paths so that paths with spaces don't break make
- Use consistent hyphenated naming, instead of sometimes camelCase and sometimes dash-case
- Speed up the tests by 2x to 1000x, depending on the IO speed, by symlinking node_modules instead of re-installing for EVERY test.

Fixes #240
Closes #267
Closes #268
Closes #269
Closes #270
Closes #271
Closes #272
Closes #273
Closes #274
Closes #275